requirements.txt
flask==2.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
